pyproject.toml
[tool.poetry]
name = "betl"
version = "0.1.0"
description = ""
authors = ["knoriega <knoriega@alum.mit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
injector = "^0.20.1"
click = "^8.1.3"
pandas = "^2.0.1"
sqlalchemy = "^2.0.13"
mypy = "^1.3.0"
pyyaml = "^6.0"
attrs = "^23.1.0"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
